+++ plugin_info/install.php
@@ -262,6 +262,33 @@ function convertBatteryStatus() {
 	jMQTT::logger('info', __("Commandes batterie définies directement sur les équipements jMQTT", __FILE__));
 }
 
+function moveCertsInDb() {
+	$oldPath = realpath(dirname(__FILE__) . '/../data/jmqtt/certs');
+	// for each brokers
+	foreach ((jMQTT::getBrokers()) as $broker) {
+		$broker->setConfiguration(jMQTT::CONF_KEY_MQTT_PROTO, boolval($broker->getConfiguration('mqttTls', 0)) ? 'mqtts' : 'mqtt');
+		$broker->setConfiguration('mqttTls', null);
+		$cert = $broker->getConfiguration('mqttTlsCaFile', '');
+		$broker->setConfiguration('mqttTlsCaFile', null);
+		if ($cert != '')
+			$broker->setConfiguration(jMQTT::CONF_KEY_MQTT_TLS_CA, file_get_contents($oldPath.'/'.$cert));
+		$cert = $broker->getConfiguration('mqttTlsClientCertFile', '');
+		$broker->setConfiguration('mqttTlsClientCertFile', null);
+		if ($cert != '')
+			$broker->setConfiguration(jMQTT::CONF_KEY_MQTT_TLS_CLI_CERT, file_get_contents($oldPath.'/'.$cert));
+		$cert = $broker->getConfiguration('mqttTlsClientKeyFile', '');
+		$broker->setConfiguration('mqttTlsClientKeyFile', null);
+		if ($cert != '')
+			$broker->setConfiguration(jMQTT::CONF_KEY_MQTT_TLS_CLI_KEY, file_get_contents($oldPath.'/'.$cert));
+		$broker->save();
+	}
+	// rm -rf $oldPath
+	foreach (array_diff(scandir($oldPath), array('.','..')) as $file) {
+		unlink($oldPath.'/'.$file);
+	}
+	rmdir($oldPath);
+	jMQTT::logger('info', __("Certificats déplacés dans la base de donnée", __FILE__));
+}
 
 function jMQTT_install() {
 	jMQTT::logger('debug', 'install.php: jMQTT_install()');
@@ -341,9 +368,14 @@ function jMQTT_update($_direct=true) {
 			convertBatteryStatus();
 			config::save(VERSION, 10, 'jMQTT');
 		}
+		// VERSION = 11
+		if ($versionFromDB < 11) {
+			moveCertsInDb();
+			config::save(VERSION, 11, 'jMQTT');
+		}
 	}
 	else
-		config::save(VERSION, 10, 'jMQTT');
+		config::save(VERSION, 11, 'jMQTT');
 }
 
 function jMQTT_remove() {